2023 10 - part 2
I did not complete this, but found several good algorithm examples. I
think my favorite is the even-odd rule. See enclosed link to wikipedia
article. I think I may could have implemented this on my own if I had
searched for an algorithm for "how to determine if a point is inside or
outside a polygon." Alas, I did not search for this.