2023 20 - part 2
Kinda got lucky here by trying the "look for a cycle count" after
realizing that a brute force was likely to take forever.

The idea is that we are looking for a time when all the inputs to "rx"
fire a high signal at once so that rx receives a low signal.

What feeds rx in our example is a single node, "kc", but more things
(four in fact) feed that. So we look for when those all fire and see how
many presses it took to get there. We then take the least common
multiple of those and get the answer.

I am not sure *how* that guarantees that only a single low pulse fires
off to this, but I was guessing that a cycle count approach would work
and viola it gave the right answer.